优化框架
IAxis
增加SetParameters方法，使用无参数构造函数实例化对象后初始化对象参数

AxisBase
修改局部变量的命名方式
增加无参数构造函数，以支持MotionControllerBase的范型定义
增加SetParameters方法，见IAxis接口

IrixiAxis
重载SetParameters方法，增加马达运动方向翻转设置的代码

IMotionController
移动MotionControllerType定义到EnumDefinition.cs文件
增加GetAxisByName函数，用于返回特定AxisName的IAxis实例

MotionControllerBase
增加范型支持，实例化时需指定其包含Axis对象的类型
增加GetAxisByName函数，详见IMotionController接口

IrixiEE0017
修改MoveProcess函数签名，将Distance更改为Steps，以免造成理解歧义

+++ Irixi_Aligner_Common/Classes/BaseClass/RealworldPositionManager.cs
@@ -127,10 +127,10 @@ namespace Irixi_Aligner_Common.Classes.BaseClass
         }
 
         /// <summary>
-        /// this method exists specially for the Luminos P6A.
-        /// For the P6A, the travel distance and max steps are given, 
-        /// but the max steps are stored in the external flash of the P6A, so this property should be re-set after
-        /// initializing the P6A.
+        /// This method exists specially for the Luminos P6A.
+        /// For the P6A, though the travel distance is defined in the profile, 
+        /// it might be different from the value stored in the P6A's flash which has the higher 
+        /// priority, so the MaxSteps should be changed to the value in flash.
         /// </summary>
         /// <param name="MaxSteps"></param>
         public void ChangeMaxSteps(int MaxSteps)
@@ -138,7 +138,7 @@ namespace Irixi_Aligner_Common.Classes.BaseClass
             this.MaxSteps = MaxSteps;
 
             // if the property was set, the related properties must be recalculated
-            this.Resolution = this.TravelDistance / (double)this.MaxSteps;
+            this.Resolution = this.TravelDistance / this.MaxSteps;
         }
 
         public override string ToString()
